+++ packages/core/src/ai-tracing/default.ts
@@ -6,18 +6,17 @@ import { MastraError } from '../error';
 import type { IMastraLogger } from '../logger';
 import { ConsoleLogger, LogLevel } from '../logger';
 import { MastraAITracing } from './base';
+import { ConsoleExporter } from './exporters';
 import type {
   AISpanType,
   AISpan,
   AISpanOptions,
-  AITracingExporter,
   AITracingInstanceConfig,
-  AITracingEvent,
   AISpanTypeMap,
   AISpanProcessor,
   AnyAISpan,
 } from './types';
-import { SamplingStrategyType, AITracingEventType } from './types';
+import { SamplingStrategyType } from './types';
 
 // ============================================================================
 // Default AISpan Implementation
@@ -332,110 +331,6 @@ export class SensitiveDataFilter implements AISpanProcessor {
 }
 
 // ============================================================================
-// Default Console Exporter
-// ============================================================================
-
-export class DefaultConsoleExporter implements AITracingExporter {
-  name = 'default-console';
-  private logger: IMastraLogger;
-
-  constructor(logger?: IMastraLogger) {
-    if (logger) {
-      this.logger = logger;
-    } else {
-      // Fallback: create a direct ConsoleLogger instance if none provided
-      this.logger = new ConsoleLogger({
-        name: 'default-console-exporter',
-        level: LogLevel.INFO, // Set to INFO so that info() calls actually log
-      });
-    }
-  }
-
-  async exportEvent(event: AITracingEvent): Promise<void> {
-    const span = event.span;
-
-    // Helper to safely stringify attributes (filtering already done by processor)
-    const formatAttributes = (attributes: any) => {
-      try {
-        return JSON.stringify(attributes, null, 2);
-      } catch (error) {
-        const errMsg = error instanceof Error ? error.message : 'Unknown formatting error';
-        return `[Unable to serialize attributes: ${errMsg}]`;
-      }
-    };
-
-    // Helper to format duration
-    const formatDuration = (startTime: Date, endTime?: Date) => {
-      if (!endTime) return 'N/A';
-      const duration = endTime.getTime() - startTime.getTime();
-      return `${duration}ms`;
-    };
-
-    switch (event.type) {
-      case AITracingEventType.SPAN_STARTED:
-        this.logger.info(`🚀 SPAN_STARTED`);
-        this.logger.info(`   Type: ${span.type}`);
-        this.logger.info(`   Name: ${span.name}`);
-        this.logger.info(`   ID: ${span.id}`);
-        this.logger.info(`   Trace ID: ${span.traceId}`);
-        if (span.input !== undefined) {
-          this.logger.info(`   Input: ${formatAttributes(span.input)}`);
-        }
-        this.logger.info(`   Attributes: ${formatAttributes(span.attributes)}`);
-        this.logger.info('─'.repeat(80));
-        break;
-
-      case AITracingEventType.SPAN_ENDED:
-        const duration = formatDuration(span.startTime, span.endTime);
-        this.logger.info(`✅ SPAN_ENDED`);
-        this.logger.info(`   Type: ${span.type}`);
-        this.logger.info(`   Name: ${span.name}`);
-        this.logger.info(`   ID: ${span.id}`);
-        this.logger.info(`   Duration: ${duration}`);
-        this.logger.info(`   Trace ID: ${span.traceId}`);
-        if (span.input !== undefined) {
-          this.logger.info(`   Input: ${formatAttributes(span.input)}`);
-        }
-        if (span.output !== undefined) {
-          this.logger.info(`   Output: ${formatAttributes(span.output)}`);
-        }
-        if (span.errorInfo) {
-          this.logger.info(`   Error: ${formatAttributes(span.errorInfo)}`);
-        }
-        this.logger.info(`   Attributes: ${formatAttributes(span.attributes)}`);
-        this.logger.info('─'.repeat(80));
-        break;
-
-      case AITracingEventType.SPAN_UPDATED:
-        this.logger.info(`📝 SPAN_UPDATED`);
-        this.logger.info(`   Type: ${span.type}`);
-        this.logger.info(`   Name: ${span.name}`);
-        this.logger.info(`   ID: ${span.id}`);
-        this.logger.info(`   Trace ID: ${span.traceId}`);
-        if (span.input !== undefined) {
-          this.logger.info(`   Input: ${formatAttributes(span.input)}`);
-        }
-        if (span.output !== undefined) {
-          this.logger.info(`   Output: ${formatAttributes(span.output)}`);
-        }
-        if (span.errorInfo) {
-          this.logger.info(`   Error: ${formatAttributes(span.errorInfo)}`);
-        }
-        this.logger.info(`   Updated Attributes: ${formatAttributes(span.attributes)}`);
-        this.logger.info('─'.repeat(80));
-        break;
-
-      default:
-        throw new Error(`Tracing event type not implemented: ${(event as any).type}`);
-    }
-  }
-
-  async shutdown(): Promise<void> {
-    this.logger.info('DefaultConsoleExporter shutdown');
-  }
-}
-
-// ============================================================================
 // Default Configuration (defined after classes to avoid circular dependencies)
 // ============================================================================
 
@@ -443,7 +338,7 @@ export const aiTracingDefaultConfig: AITracingInstanceConfig = {
   serviceName: 'mastra-ai-service',
   instanceName: 'default',
   sampling: { type: SamplingStrategyType.ALWAYS },
-  exporters: [new DefaultConsoleExporter()], // Uses its own fallback logger
+  exporters: [new ConsoleExporter()],
   processors: [new SensitiveDataFilter()],
 };
