+++ packages/core/src/ai-tracing/exporters/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Mastra AI Tracing Exporters
+ */
+
+// Core types and interfaces
+export * from './console';
